requirements.txt
python-telegram-bot==20.7
google-generativeai==0.8.3
openai==1.54.0
